Add a more user-friendly handling of the case where the nix gets removed from /etc/zshrc
This change adds more informations to the terminal when macos update removes the nix and home manager from the /etc/zshrc. It tells the user how to recover from this.

It also updates the restore-nix.sh file to only write the config to /etc/zshrc when nix is actually not present. This prevents from accidentally writing nix multiple times into that config.